Added initial tests.
